Fix a couple of logging messages
For some reason when reading the logs you can't use the multiple
argument form of `console.log`. When you do that the Pebble logging
output doesn't print the second argument. It works fine in the Emulator
but not when connecting to a real device. :rage3:

I'll just concatenate them for now.