Brawl Scheduler should use Daylight Savings Times
I was an hour off because I didn't realise that in the US and Europe
they're currently on PDT (instead of PST) and CEST (instead of CDT).

I was basing my times off the http://hearthstone.gamepedia.com/Tavern_Brawl#Schedule
site. It's actually hard to find a source of truth for this stuff.